+++ src/main/java/models/IVector.java
@@ -4,5 +4,5 @@ import java.util.Vector;
 
 public interface IVector {
 
-    Vector<Integer> vector();
+    Vector<Double> vector();
 }
